Исправит наименования переменных, файлов
