+++ src/main.js
@@ -1,15 +1,14 @@
 import {render} from './render.js';
 import TripInfoView from './view/trip-info-view.js';
-import FilterView from './view/filters-view.js';
-import MainPresenter from './presenter/presenter.js';
+import FiltersView from './view/filters-view.js';
+import MainPresenter from './presenter/main-presenter.js';
 
-const tripMain = document.querySelector('.trip-main');
-const tripControlsFilters = tripMain.querySelector('.trip-controls__filters');
-//const tripEvents = document.querySelector('.trip-events');
-//const mainPresenter = new MainPresenter({boardContainer: tripEvents});
+const tripMainElement = document.querySelector('.trip-main');
+const tripControlsFiltersElement = tripMainElement.querySelector('.trip-controls__filters');
+const tripEvents = document.querySelector('.trip-events');
+
+render(new TripInfoView(), tripMainElement, 'AFTERBEGIN');
+render(new FiltersView(), tripControlsFiltersElement);
+
 const mainPresenter = new MainPresenter();
-
-render(new TripInfoView(), tripMain, 'AFTERBEGIN');
-render(new FilterView(), tripControlsFilters);
-
-mainPresenter.init();
+mainPresenter.init(tripEvents);
